+++ NWP_projekt/General.h
@@ -15,7 +15,7 @@ public:
 	~General();
 	void InitializeGame(CClientDC* pDC, RECT cr);
 	void PlayMove(CDC* pDC, RECT cr);
-	void PlacePiece(CClientDC* pDC, RECT curr_field, RECT cr, int piece_color, int piece_type);
+	void PlacePiece(CClientDC* pDC, RECT curr_field, RECT cr, double piece_color, double piece_type);
 };
